added more explicit docs, changed Kind from enum to struct backed by an enum
